Popularity calculation cron job. Fixed problem where logging in with wrong credentails showed a "session expired" message. Fixed problem where User Profile Page wasn't including the current user's host groups. Updated "autocomplete" attribute to "autoComplete". Set a character count limit for Show Builder comments.
