+++ package.json
@@ -19,6 +19,7 @@
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "build": "NPM_CONFIG_PRODUCTION=false npm install --prefix client && npm run build --prefix client",
     "heroku-postbuild": "npm run build",
+    "calculatepopularity": "node server/v1/calculatePopularity",
     "seed": "node server/v1/seedDB",
     "seed:test": "cross-env NODE_ENV=test npm run seed",
     "seed:dev": "cross-env NODE_ENV=development npm run seed",
