+++ server/v1/controllers/library/index.js
@@ -1,3 +1,4 @@
+const calculatePopularity = require('./calculatePopularity');
 const create = require('./create');
 const createMany = require('./createMany');
 const findAll = require('./findAll');
@@ -7,6 +8,7 @@ const search = require('./search');
 const update = require('./update');
 
 module.exports = {
+  calculatePopularity,
   create,
   createMany,
   findAll,
